Merged build instructions for overenstr into main build script. (Ticket #2, #5)
For unknown reasons (possibly a late readdition after the build script had been unified), the build instructions for overenstr resided in a different script than other programs. 
These are now in the main build script, and the separate script remove. The install script already contained the appropriate command and did not need to be altered.

(This commit is a merger from the theia_unify_dev branch that made these changes.)