Support compilation on Theia
Update build script and module file for compilation on Theia. To avoid
conflicts with tide, the first two letters of the hostname are now also
used to determine which machine, and the check for Theia is first.

Modules loaded on Theia follow the version for Cray as much as possible,
with the following differences due to inavailability for now (excluding
architecture-specific modules):
png is 1.2.44 instead of 1.2.49
z is 1.2.6 instead of 1.2.7
there is no iobuf module

All executables are compiled, but there is not yet a guarantee they
function properly.

Change-Id: Ibca1caa2d5d4d2e472e9d9c14f8ec3f0d1c2b703
Refs: #44962