Allow workflow file to work with rocoto_viewer
Previously rocoto_viewer could not parse the external entites we use in
the workflow XML, forcing us to use the inferior rocotostat to view job
status. rocoto_viewer has separately been updated to allow for external
entities, but requires the full path of the entities in the XML file.
This change uses @[GEFS_ROCOTO] in gefs.xml.in to generate full path
names in the gefs.xml workflow when the file is generated.

Change-Id: I7b0e0685cbd641de8d4c96d3e9b9a3396679a8c2