Fix chgres task in init_separate
As of global_shared.v15.0.0, chgres was updated to produce FV3
cube-sphere faces. The old chgres scripts and executables were renamed
with the suffix _GSM. To allow backwards compatibility to v14.1,
chgresush and CHGRESEXEC in JGEFS_INIT_SEPARATE now allow for
pre-existing definitions, which were added to setbase.

The definition of APRUNC was also incorrect and needed to be updated.

Also increased the vmem request further for enkf_track and
init_separate to avoid jobs being killed by the scheduler. It is unknown
why Theia seems to have much higher vmem requirements for jobs. These
values (and walltime) will need to be reassessed after a fully
functioning version theia is complete.

Change-Id: I9df597559e4dbd256e9d8c7883ad12493a359b2c
Refs: #44962