Add Theia support for forecast job
While this commit adds Theia support for the forecast job, the output is
not in the expected place. The global forecast script
(exglobal_fcst_nemsfv3gfs.sh)  inside the FV3-GFS packagespecifies a
certain directory stucture under $ROTDIR when running a non-zero member
and this does not have an override. When Bing built the system for
WCOSS-Cray, he slightly modified the global forecast script. However,
this is not a good permanent solution. A future GEFS commit will change
the ROTDIR and symbolic links so GFS can put the output in the location
it wants but still give us files in the correct location.

Name of $DIAG_TABLE and $FIELD_TABLE and location of forecast executable
are updated to reflect a more recent version of GFS. To keep existing
compatability on WCOSS-Cray, this is done using a default-and-override
system. Later the default will need to be updated (probably as part of
the update for the flat GFS structure).

APRUN was hardcoded in gefs_fcst.parm and this was short-circuiting the
previously set value.

The name of the environment variable pointing to the FV3 fix data was
changed from $FIX_FV3 to $FIXfv3 and the appropriate directory is
currently fix_fv3_gmted2010. Add the new variable into the job script
without removing the old one (for now) to maintain existing
compatability.

As with previous jobs, vmem was increased to avoid a halt from the job
scheduler. Correct value will be determined later.

Change-Id: I57eb04397041622123a38012bce21ab89e8d94c1
Refs: #44962