Allow GEFS to use standard global forecast script
The standard global forecast script (exglobal_fcst_nemsfv3gfs.sh) places
output in a specific directory ($memdir) based on $ROTDIR:
memdir=$ROTDIR/enkf.$PDY/$cyc/$mem. This directory structure does not
coincide with the GEFS COM directory structure and there is no way to
directly override this setting. When developing the FV3-GEFS system,
Bing modified the global forecast script to use a different directory,
but this is not a permanent solution.

To avoid requesting a change to the global script, I changed the
definition of $ROTDIR in the GEFS forecast job and created a symbolic
link to the appropriate GEFS COM location where FV3 expects $memdir
to be.

Two additional modifications Bing made to the global script related to
stochastic physics will be addressed in a separate commit.

Change-Id: I86af7ecad502e8d738c588829933ea233bc4ffa1
Refs: #44962