Add Theia support for JGEFS_NCEPPOST
The UPP script exgfs_nceppost.sh.ecf fails if the error script cannot be
found (such as when it is set to the WCOSS value of err_chk). This
required the setting of ERRSCRIPT to an value of " " and modifying
jobs/JGEFS_NCEPPOST to accept an override.

The number of nodes was increased to 3 for Theia as the job would fail
(likely due to memory limits) on fewer nodes. Increased the number of
threads to utilize all of the nodes.

Change-Id: Iad882c5313c75dbb51617d6c5d686736d9e150d9
Refs: #44962