Create workaround for modified global fcst scripts
A lot of the early testing for FV3 is being done with a modified global
forecast script that uses a different memdir (output directory) location
than the standard script. This creates problems when creating symlinks
for the model output in COM. As a workaround, I have added a memdir
template that allows users to specify the path to the memdir.

To avoid disruption to existing workflows for testing already underway,
the modified memdir path location is currently the default ($ROTDIR). In
order to use the unaltered global forecast script, set
memdir_template='$ROTDIR/enkf.$CDUMP.$PDY/$cyc' (must use single quotes
to delay variable expansion).

Change-Id: Ifd639efb9ea258443231d647854030ed592d11b5
Refs: #44962