Replace hardcoded cleanup directories
Replaces the variables that were hard-coded in the cleanup scripts with
variables that are set in the XML workflow and passed to the python
scripts via environment variables. There are four new entities, which
are turned into environment variables by env_var.ent:

KEEP_DIR:
	Directory to place output we wish to keep on spinning disk (usually
	noscrub)

DIRS_TO_KEEP:
	A comma-separated list of directories (e.g., pgrb2ap5) to keep on
	spinning disk. These will be copied to KEEP_DIR.

HPSS_DIR:
	Location on HPSS for archived output.

DIRS_TO_ARCHIVE:
    A comma-separated list of directories (e.g.,cpgrb2ap5) to archive
	on HPSS. These will be placed in a tar file in HPSS_DIR.

Additional entries are added to user.conf_prototype to set these
variables. A new variable RUN_CLEANUP is also added which adds/removes
the three cleanup jobs to the workflow. gefs.xml.in and env_vars.ent are
modified to pass these values along to the scripts.

Documentation and informational prints are also added to the cleanup
scripts.

Change-Id: I97dc7a857882d5dd9804b01e604b61ddc4024a00
Refs: #46726