Fix queue and python issues with cleanup jobs on Cray
Python scripts used for cleanup had to be changed in order to make them compatable
with Python 2.7, as Python 3.x is not available yet.

Additional changes to the archive task description were made to function properly
with the Cray transfer queue.

The hpss module had to be loaded to properly interact with HPSS in the archive job.

Change-Id: I229e5f2500ed0833cc309a6d55f06532240ac994
Refs: #46726