Keep global forecast script from overriding SP seeds as default
The global forecast script will set the stochastic physics seeds if
SET_STP_SEED="YES". This is mostly a problem for SPPT, where we use five
scales instead of one, and thus need to use a list of five seeds. This
changes the setting so the global script does not overwrite our seeds,
and also changes our [high-res] seeds to the values used in Bing's
modified global forecast script.

Change-Id: I5ba4c1481c489c380db94f4344599862aa3c910e