Change wgrib2 version on Theia
The default version of wgrib2 on Theia (v0.1.9.5.1) sometimes fails
during interpolation. The newer version (v0.2.0.6c) seems to correct the
issue.

In addition to changing the module being loaded, setbase was updated to
use the wgrib2 that is already in the PATH (added by the module) rather
than pointing to a specific version. This means future updates will only
need to update the module and nothing else.

Change-Id: Id1607397d034c7c1b3c021575f3f54befb431ca1