Add a check for errors from wgrib2 during interpolation
The prdgen job was not checking the return code of wgrib2 when
interpolating. Instead only the presence of the grib file was checked.
This can result in an incomplete file without notifying the user if
wgrib fails after producing a portion of the output file. This change
adds a check of the return code after the interpolation.

Change-Id: Ie079d28dbe494d0b154470e3191aa66bb20199ad
Refs: #45912