Correct scheduler info for IBM
MPI jobs on WCOSS-IBM require '-a poe' be included in the job submission
in order to run properly. This is added via a native tag to the task
specification generated by GEFS_For_Tasks.py.

Jobs also need to not have the memory setting specified, so memory lines
are commented out of the ibm default configuation.

Change-Id: Ic56e3adc95d8d12dcdc62f8c02ad941edff629f9
Refs: #45813