Add WCOSS IBM support for fv3chgrs and forecast jobs
The number of cores used for fv3chgrs is incresed to 2 in order to have memory
for the control member.

The FV3 task geometry is updated in order to allow the forecast to run within
the wallclock limit.

Additionally, typos in specifying the queue for post-forecast jobs in the
user_wcoss_ibm.conf file are corrected.

Change-Id: I9bd18218b4eb5ff556667ceaca4b48695197bf31
Refs: #45813