Fix prdgen and archive jobs for WCOSS IBM
prdgen now requires 6 threads, so both the job request and bin script are
updated to reflect this change.

The name of the transfer queue on WCOSS IBM is just 'transfer'. Additionally,
jobs in the transfer queue require affinity and memory to be set and poe to not
be set, so the resource and native statements for transfer jobs (currently just
archive) are updated.

Change-Id: If8629d42195d36d9efe86de54450a4911ca59bff
Refs: #45813