Enclose gefs_dev.parm values in quotations
The values in assignments produced by the python script for
gefs_dev.parm are now enclosed in quotations. This both helps clarity
and is necessary for lists that uses spaces (such as for
PRDGEN_STREAMS). This change avoids the need to use special escaping in
the python configuration script (e.g. user_full.conf) to set
PRDGEN_STREAMS.

Change-Id: I7fdaa50693533bce7a1ef82524228d9915a6e23b
Refs: #58765