Merge prdgen_gfs into prdgen script and reenable makepgrb2b
Previously, prdgen_gfs was used to create products from the
deterministic GFS on the grids used for GEFS. This led to a lot of
duplication of code with the normal prdgen job, which created code
divergence as sometimes one would be updated without similar changes to
the other. This new version uses the same scripts for prdgen and
prdgen_gfs, using alternate settings when $RUNMEM = "gfs" where
requried.

Also included in this commit are changes to obey makepgrb2b. At some
point this capability had been removed from the prdgen scripts. This was
added back in at this time because 'b' files are not generated for GFS,
so it was logical to restore this capability and use it to control
pgrb2b creation for GFS files.

Change-Id: I2f2eff89894006ae56e24501cac99afb26680cd7
Refs: #58765