Allow new streams to be defined in conf file
This commit allows the definition of new streams by adding their
settings into the conf file. To accomodate this capability, several
changes are made due to the peculiarities in passing arrays between bash
script. These mostly center around running scripts further down in the
same shell.

GEFS_Parm.py is changed to always define the PRDGEN variables as
associative arrays at the begining of the gefs_dev.parm file. This must
be done before settings passed via the conf file set values for stream
keys farther down the gefs_dev.parm file.

gefs_pre_job.sh is updated to no longer use 'exec' to run the bin
script passed in as an argument. This allows the prdgen jobs to be run
in the same shell by using the dot operator while still allowing python
scripts to run normally as well.

GEFS_XML_For_Tasks.py is updated to add the dot operator to the command
for all prdgen tasks.

gefs_prdgen.parm is updated to check for the existence of the
associative arrays before defining them (and overwriting values that
have been set). (Also, a missing GRID value for 1-deg was added).

Because the j-job is now running in the same shell as the python bin
script, set -x is being passed through. This causes the script to fail
if rm attempts to remove a file that does not exist. To avoid this, the
force (-f) switch is added to rm commands in JGEFS_PRDGEN.

Error checking is also added to JGEFS_PRDGEN to ensure all necessary
settings are defined for each stream.

Change-Id: I542e5c1a51173e8e25831729f4fb75c587411a68
Refs: #58765