Correct module issue on Cray
Modules were not being purged on Cray systems at the beginning of a job. This
causes issues when modules the user loads on startup prevented different
versions from being loaded. Modules are now purged, and a couple of required
modules that are normally added automatically are added to restore them after
the purge.

Change-Id: I843798bc4725931fb80be12ff57f00e12807017f
Refs: #58765