Switch scheduler to slurm on Theia
Switches the scheduled from moab to slurm on Theia. RDHPCS is
transitioning all their HPC systems from moab to slurm, necessitating
the update.

This update completely removes the capability to use moab on
theia. Also, in order to use this version on theia, you must use rocoto
1.3.0-RC3 or later (module load rocoto/1.3.0-RC3).

To facilitate the transition, many settings common to all theia bin/
scripts were abstracted up to a bin/theia/common.sh script. For now,
this script is loaded by gefs_load_modules.inc, since that script
selectively executes code depending on machine. In the future, these
will be rearranged.

Additionally, formerly hardcoded values for several cpu geometry
settings were replaced with values reported by the scheduler. It may be
possible to add additional settings in the future. These replacements
will be made for other machines in the future as well. This supports
issue #46734.

To facilitate the new separation between QOS and partition in slurm
rather than just a queue, a new partition setting is recognized by the
python build system. For the archive job (the only job currently using
the transfer queue), the queue is changed to the normal &CUE2RUN;, and
the new archive_partition is set to the value of &TRANSFER_QUEUE;.

Job output files are updated to use %j instead of $PBS_JOBID as
specified by slurm.

Change-Id: I07f9798a324e8cf915338ff7588ed1275146f264
Refs: #59842
Refs: #46734