Update forecast job to use exglobal_fcst in global-workflow
Changes the GEFS forecast job to use the exglobal_fcst script that is
located in the master global-workflow instead of a custom GEFS version
located in the GEFS workflow. This is accomplished by making the $DATA
directory the ROTDIR and creating the expected directory structure using
symbolic links to the associated places in the GEFS directory structure.

Additionally, the name of the forecast executable is updated to
global_fv3gfs.x to match the executable generated by global-workflow
now.

Change-Id: I18dcf224c8c5eb494ed88f40d666a2c75b227738
Refs: #62882