Update forecast job to run combined FV3-WW3 model
Makes necessary changes to forecast job to run using the coupled NEMS
app.

New settings in user_full.conf control whether Wavewatch 3 should run
alongside FV3. cplwav should be set to '.true.' to run WW3 together with
FV3. npe_wav determines how many threads will be assigned to WW3 (added
to the normal npe calculation for FV3).

Wave prep/post jobs do not yet have a "quick run" option (RUN_ setting),
you must specify a full task list to run wave jobs.

Additionally, the default ptmp location for Theia is updated to be
within the ensemble project space rather than stmp4 due to space
constraints.

Technical notes:

In JGEFS_FORECAST_FV3, symlinks are created for both the wave input and
output files to the actual locations using COMINwave and COMOUTWW3.
COMINwave and COMOUTWW3 are then changed to the input directory. This
was necessary because many of the files are named differently (often
including a member id) than expected by exglobal. Since MEMBER is set to
npert+1 for the control, a new WAV_MEMBER variable is created in order
to properly produce wave file names using '00' for the control.

In gefs_fcst.parm, new variables for use when the wave ensemble is to be
run alongside are generated, primarily creation of the PET list for each
component. Additionally, the existing calculation of NTASKS/npe was
streamlined a bit.

In GEFS_XML_For_Tasks.py, the calculation of dependency for
forecast_high now adds gwes_prep as a dependency if the job is in the
task list. The resource request is also increased by npe_wav if coupling
to ww3 is to be used (cplwav == ".true.").

Conf settings for wave jobs were copied into the defaults for each
machine. Existing copy in user_full.conf were commented out, following
standard practice with the other jobs.

The location of the WW3 executables is changed to the version in the
coupled app rather than links in the local GEFS workflow.

Change-Id: I122e24bd898c91760fba2938650601546c54b26b
Refs: #58418