Add wave jobs to conf task control
A new setting to the simple task control, RUN_WAVE_PREP, was created for
use in the conf file, so task lists no longer need to be set explicitly.
When set to "YES", gwes_init and gwes_prep will run.

cplwav (used to control whethern the forecast job is coupled with WW3)
no longer needs to be set explilcitly in the conf file in most cases. If
the variable is not set, it will be set to ".true." if thegwes_prep job
is in the job flow (for instance, by setting RUN_WAVE_PREP  to "YES")
and ".false." otherwise. If cplwav is already set by the conf file, the
setting is preserved.

Wave post (gwes_post) will automatically run after the forecast job if
cplwav is ".true.". This matches ncep_post, which runs if the forecast
runs and does not have a separate control.

Additionally, some minor bugs related to the wave executable path and
the order of parm file calls and working directory creation in the wave
jobs were corrected.

Change-Id: Ie1e20edae091aa23a11d0dfda8557772ce7b4855
Refs: #58418