Fix issues with automatically determining and writing cplwav
The previous attempt to automatically set cplwav fails for a few
different reasons, which are now corrected. Determination of cplwav is
now done in run_to_get_all, so it is available to both GEFS_Parm and
GEFS_XML_For_Tasks (it is now also determined by presence of the
gwes_prep task instead of RUN_WAVE_PREP, but this wasn't required). In
order to ensure cplwav is written to gefs_dev.parm when it is determined
automatically rather than specified in user_conf, it is added to the
mandatory write list (lstBaseParm). Finally, cplwav is required to be in
lowercase, so the write step is modified to not force uppercase when
writing cplwav.

Change-Id: I0df2cf29d45b0e21e5ae15957d2daae5ea98f96c
Refs: #58418