Add WCOSS-Dell support for wave jobs
The needed rocoto run scripts are added to run the wave tasks on
WCOSS-Dell.

As part of this addition, settings common to all jobs are moved to a
common.sh script, similar to as was done for Theia. This allows any
changes to be made once, rather than in every script.

Change-Id: I0bc9d99f3af66faa287ca8a3b65a4b53caf36275
Refs: #58418