Update default wave mpi commands
The default wave_mpmd command set in the wave j-jobs was different than
what the wave scripts expect. Previously, j-jobs gave the full mpmd
command (similar to what GEFS does), but the wave scripts generate the
command by piecing together wavepmexec and wave_mpmd along with the
number of tasks. Updated wave_mpmd and added wavempexec to the wave
j-jobs for that format.

Change-Id: Ia29683689389bf2d88968cc91fde1e72055dd37f
Refs: #58418