Update post job for chemistry
To support the inclusion of chemistry tracers, new flat files were
added to the ncep_post repository. Setbase is temporarily updated to
point to a development copy of ncep_post that includes the new flat
files, as well as point to those new files. Current testing is with GFS
versions of these flat files. GEFS version will be created later
(possibly with separate chemistry files from the GEFS meteorology
data). The settings will be moved into the nceppost j-job or a new chem
nceppost j-job in the future.

Additionally, the inclusion of chemistry to post requires fixed aerosol
optial data. These files are now linked by the j-job when ESRL_CHEM=YES,
though this will likely be updated to use cpl_chem in the future. If
chemistry post is processed separately, these links would be moved to
the new chem-post j-job.

Change-Id: If08a7b66eec4c1f65769ddf5230c0bcd80ad60be
Refs: #62104