Update to use chgres in ufs_utils
The location of chgres in the global workflow has changed. It is now
located within the new ufs_utils.

Unfortunately, global_chgres_driver.sh hasn't been fully updated to
reflect that change, so for now we have to override HOMEgfs to point to
the ufs_utils directory and set FIXfv3 and FIXam beforehand.

Change-Id: I22351f1d167109bd7d17de2fbbb25114561a99b0