Update and cleanup setbase
The HOMEDIR location is updated for all machines except WCOSS-IBM to
the current version of the global-workflow branch for the coupled app.

The tracker location is updated to v2.1.2 on machines where it is
currently available (WCOSS-Cray and WCOSS-Dell).

Cleaned up and removed a lot of settings that are no longer needed.
Remaining settings for using real-time GFS inputs updated to use the
preferred compath.py script.

Also added a placeholder block for Hera, which will also stop Hera from
being misidentified as Theia before the port has been completed.

Change-Id: Ia880c99c9114896a8cadf9ca5960d5f37ddea84f
Refs: #58418, #57673, #67587