Add capability to run wave on Cray
Updated mpi/mpmd ALPS settings for the wave jobs so they use the correct
command (aprun).

Also had to add an EXECcode to setbase for Cray to provide the correct
location of wave executables. Since this is needed for all machines and
is the same, in the future this will need to be moved out of setbase to
somewhere more appropriate.

Change-Id: Ic5ff2d1dd3367d3d18605bbf7343e29171f2c859
Ref: #58418