Condense python task scripts for each machine to a single task script
With Python 3 now used on all machines, only one script is needed for
each task. A new py directory is created in the bin directory. The Dell
version of the python scripts are moved to the new directory and other
versions are removed.

Additionally, cleanup task names are updated with an _atm suffix in
preparation for upcoming similar tasks for wave data. Resource blocks
and code for new _wave versions are added, but task scripts have not
been created yet.

Change-Id: I454eae7b3900355cf936d3db2550446acbebf219
Refs: #58418