Add aersol forecast job
Adds a forecast aerosol job. When used in the workflow, wave coupling
is automatically turned off for that job. RUNMEM is set to geaer. There
is currently no independent settings for the FV3 layout, so whichever
settings (and the associated CPU request) are used for the atmospheric
portion of other members is used. Wall clock can be defined independently
however. The forecast_high bin script is reused.

A new parameter, fhmax_aer controls how long the aerosol member runs.
The default value is 120. When RUNMEM is geaer cplchm is automatically
set to true by gefs_fcst.parm. When cplchm is true (either because
RUNMEM is geaer or it was set true beforehand), fhmax and fhmaxh are set
to fhmax_aer and other settings necessary for chemistry coupling are set.
Additionally, the restart_frequency is set to gefs_cych to produce the
appropriate restart files for aerosol tracers. This final setting is
temporary until the capability to produce restart files at two cadences
is added.

In the forecast j-job, a new section links the first tracer restart files
to the COM directory if cplchm is true, to be used to initialize aerosols
for the next cycle. To prevent unnecessary symlinks for output files for
aerosol jobs (when fhmax_aer is used), the high-frequency loop now checks
both FMAX_HF and FHMAX. Additionally, some unused settings were removed.

Change-Id: Icce1a4b40eab2bf57acf7635ab4571d9728ea4d4
Refs: #62104