Fix bug with aerosols turning off wave coupling
Having an aerosol_forecast was turning off wave coupling for all
members instead of just for the aerosol task. Removed the offending
code from the rocoto side and added code in the parm file to turn
off wave coupling when chem coupling is on, which is typically only
for the aerosol member.

Change-Id: Ib6b107a40582c28ec90e3263571d3ebf5292ad93
Refs: #62104