Fix bugs related to aerosol cycling
NetCDF operators module was not loaded, which cause an error in init_aerosol
because ncks could not be found. This resulted in tracers being unmodified
(zero) in init files.

Setting for the additional restart output time added and restart interval
returned to its original value.

Forecast was using the wrong offset for aerosol cycling to create symlinks for
restart files for cycling.

Increased aerosol post reqsources from 4 to 5 nodes due to malloc error.

Change-Id: Ib2d615acdfa398136932cc21e168f9062afbb1ea
Refs: #62104