Add warm-start capability for aerosol member and make default
Adds the ability to warm-start FV3 in aerosol cycling rather than
cold start and makes this the default method. If init_aerosol runs
and is unable to generate an increment file (usually because the
restart files are missing), it will fail over to a cold-start option.
This will technically try to use the restart files tracers, but since
failure is generally because restart files don't exist, in reality
this generally means using the normal cold-start initialization (same
as control) with zeroed tracers. In a development setting, this procedure
should only occur for the first cycle, and prerequisites are set to wait
for the restart files to be available before submitting the init_aerosol
job.

The new initialization mode requires the linking of additional restart
files to COM. Right now we might be a little overzealous with what's
being saved (for instance, I'm not sure sfc_data files need to be saved
because the analysis fields are being used); this can be revisited later.

The old cold-start aerosol cycling method is still available, but it is
currently considered suspect as there are no adjustments to conserve mass
when the tracers from the restart files are copied into the new inital
files. This setting can be accessed by setting AEROSOL_INIT_TYPE to "cold".

Change-Id: I033600f7f1e338e989f68471f4b13c21f81b1eb7
Refs: 62104