Port aerosol jobs to Cray (mostly)
Adds scripts necessary to run aerosol jobs on WCOSS-Cray. However,
because the forecast and post jobs use different resources than the
normal members and Cray does not have variable definitions (in this
branch) for aprun commands, those jobs do not run properly. Since the
variable definitions have already been implemented in master, those
modifications are not duplicated here at this time. The commit was
tested by manual modifications to the necessary bin files, but those
changes are not comitted.

In order to obtain the necessary memory for the prep_emissions job,
the script has to be run using aprun even though it is a serial job.
Therefore, a $gefsmpexec was added to the script call in the j-job.

Change-Id: I92cb14b33423a27cef6b013347b4cb72a1940ae8
Refs: #62104