Fix post reproducibility and some post error-checking
Previous setting using Open MP is known to be unable to reproduce
post output, so the post job is changed to use OMP_NUM_THREAD=1.
The node request remains unchanged for now due to concerns about
memory and the need for a quick change; the exact resource request
should be revisited later.

There was a problem in the post script with comparing certain
numbers with leading zeros in mathmatical context because they were
being treated as octals. Forcing a conversion to base 10 corrects
this issue.

The return code was not being checked properly in the j-job. This is
corrected, but a more thorough examination of the error checking is needed.

Change-Id: I8d71ea35938ee4ecbf23c1db8db76a877eea8ea4
Refs: #62104