Detect emission type for forecast
While prep emissions is designed to fail-over to MODIS, WF_ABBA, and
then climatology if GBBEPx isn't available, the emission type still
had to be set manually for the forecast. Now, the emission type is
detected automatically by looking for the presence of plumefrp.dat
files, which are only present for GBBEPx.

Change-Id: I0751d71ea0956d8c553aac1608388f07dc778abe
Refs: #62104