Fix location of restart directory for aerosol files
Due to recent changes elsewhere, the symlinks for the aerosol restart
files were being created in the wrong location, leading to them not
being picked up by the forecast for warm start on subsequent cycles.

Also added a condition so that stochastic restart files are only copied
if stochastic physics is used.

Change-Id: Ica7682b9cf4bfe35e1d17fd3330aca1ac025383a
Refs: #62104