Fix mismatch in task/job name for wave_stat
The new wave stats job is named wave_stat, but the rocoto system
was using wave_stats, leading to the bin file not being found.
Updated all references to be consistent.

Also fixed a minor issue with setpdy in wave stat. The j-job still
needs to be more thoroughly updated to be consistent with other jobs.

Change-Id: Ia661aca7f6e3e5468492071ee30e8ff6f7eed898
Refs: #58418