Fix error handling for dangling symlinks in copytree
The previous method for catching errors from copytree was not
sufficient. copytree throws a special shutil.Error exception rather
than a FileNotFoundError, which is a list of tuple strings. Now the
script checks whether the exception strings in the tuple all indicate
a file not found error. If so, the error is ignored. If another
exception is present, the exception is raised.

Change-Id: I37b7e922f15375137d4e3d2908d91b0253aaafae
Refs: #62104