Improve prdgen customizability
New settings are added to specify the variable list to use when
producing pgrb2a and prgb2b files. There are separate settings for
A and B files, as well as for f00 and other hours. This allows
different parm files for different streams, including elimination
of some hardcoded settings from the scripts.

Additionally, B files are now automatically turned off if
PRDGEN_B_DIR is empty or undefined for the stream.

Change-Id: Ifbcc08fddb11b5c66ae26c30d3587a60018580dc
Refs: #57673