Fix bug with line length limit for prdgen
After the addition of the parmlists to the command line in the mpmd
script, the lines were sometimes too long if a lot of hours were
requested for a stream. Now, all of the settings are written to an
input file which is passed to the exscript instead. The input file
is just a bash file with variable assignments for that stream. This
file is sourced near the beginning of the exscript.

Change-Id: I8626fe8c37f97893ee49042a62221ada3b7fbd9b
Ref: #64583, #76266