Update post for new COM structure
Update the post job for the new COM structure. Like the forecast
job, the parm sourcing had to be moved earlier to allow us to
determine COMPONENT based off of cplchm.

The changes to the exscript are primarily to account for the
different contents of COMIN/COMOUT (which now include the cycle in
addition to the component). Some commented dead code was also
removed to clean things up.

Change-Id: I23863610f40b2e9d60e6980bdf17241827a4c96f
Refs: #63941, #76266