Update prdgen for new COM structure
Update the prdgen job for the new COM structure. Like the previous
job, the parm sourcing had to be moved earlier to allow us to
determine COMPONENT based off of cplchm.

The changes to the exscript are primarily to account for the
different contents of COMIN/COMOUT (which now include the cycle in
addition to the component).

Change-Id: I60fa1365266e62a580816e95e2781d8eb225bce8
Refs: #63941, #76266