Move j-jobs unused in operations
All of the j-jobs not used for operations are moved inside the
rocoto directory under rocoto/bin/sh. All of the run scripts for
these jobs in rocoto/bin/<machine>/ have been updated to point
to the new location.

This change should reduce confusion and mirrors the existing py
directory for dev-only python scripts.

Change-Id: Ia4ac0e1c5c685057e960f65aff52947e077f1ee7
Refs: #63941, #76266