Fix issue with COMIN_WAV_ICE using PDY twice
A recent change to the wave prep j-job altered COMIN_WAV_ICE so that
gefs.$PDY/$cyc was appended to whatever had been defined before.
This caused gefs.$PDY/$cyc to appear twice. Change this by moving
the $PDY addition inside the } so that it is only appended when
needed to the result from compath.py in the default situation.

Change-Id: Ic8688de90f4bb7b9f40663a5d677ce621979e6f9
Refs: #63941, #76266