Update prep emissions for better reproducibility
Previously, prep emissions would always use the most recent GBBEPX
smoke emissions available. However, since the current day's files
are generally not available for the 00z cycle, this causes an issue
where which file is used for 00z depends on how early it is run.
Now GBBEPX_DATA_DIR is broken up into two variables for today and
yesterday. The yesterday version is used for 00z, while today is
used for other cycles.

This has an associated update to global-workflow to pick up the
change to gsd prep emissions.

Change-Id: I3065772c32f24145dd335fa8e0bc6a6be73f5c99
Refs: #62104, #63941, #76266