Fix undefined variables and standardize
Fixed all instances where undefined variables were being used. In
most cases this just meant deprecated variables were removed from
echos or variables names, or ensuring files exist before deleting
or listing them. This process also turned up several typos or
other errors where code was not updated for some new feature yet
continued to work or failed in an unnoticable way.

Notes on non-trivial modifications follow

All ecflow_client statements were wrapped in SENDECF checks so
they are only called if ecflow is being used.

All ERR were changed to err so err_chk operates correctly.

In forecast, only copy the stoch_ini file if stochastic physics
is being used.

In atm post, there was a typo where FHMAX_HF was used instead of
FHMAXHF, so the start hour was always FHOUTLF after fhmaxh.
Since FHMAXHF is usually less than fhmaxh and the correct branch
was taken anyway, this was unnoticed.

Also in atm post, the determination of the perturbation type for
the GRIB2 headers had not been updated for the aerosol member.
This section was rewritten to consider the geaer and remove
unused code for negative perturbations.

In atm prdgen, the stream name was not being passed from the jjob
to the script. Added it as a command line argument instead of the
input file so it could be used immediately by the script for PS4.

In enspost, added a check so we only run wgrib2 if $pgtem exists.

Change-Id: I27d167f1c256c6cb140540659a203c307e15a410
Refs: #63941, #76266