Clean j-job and error messages for atmos_prep
Remakes the j-job to NCO standards. Places where jid ($$) were
used are replaced with $LSB_JOBID. A lot of code from the j-job
is moved down to a new script-level script in order to remove it
from the j-job. Some settings are pushed down to the parm file.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266