Clean j-job and error messages for init_recenter
Remakes the j-job to NCO standards. Places where jid ($$) were
used are replaced with $LSB_JOBID. Some code is pushed down to
the script-level script. Some settings are pushed down to the
parm file.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266