Rearrange locations for initialization jobs
In order to meet NCO standards about obeying $SENDCOM, the
locations of initialization files has been juggled around. Instead
of writing directly to $COM and copying back to $GESIN, files are
created in $GESIN and copied to $COM if appropriate.

This requires some knock-on changes. Recenter was trying to use the
same file as input and output. Instead, the files from prep are now
written to a new directory in NWGES named enkf. The final init files
from recenter are written to another new directory named init. The
files in init are then copied when appropriate.

During these changes, some deficiency issues with the recenter
python script arose. First, because cfp does not always return an
error when mpmd jobs fail, the whole main function is now wrapped
in a try block to catch any uncaught exceptions and call err_chk,
which will abort the job. Second, a race condition existed where
multiple jobs could try to create a directory at the same time,
resulting in failures because the directory existed. Now, instead
of checking if the directory exists (which caused the race
condition), the directories are just made and FileExistsError
exceptions are ignored. Finally, a system call to cp was replaced
with the Python copy method and partial is used to automatically
flush print statements instead of explictly calling it every time.

Refs: #63941, #76266