Clean j-job and error messages for chem_init
Remakes the j-job to NCO standards. Places where jid ($$) were
used are replaced with $LSB_JOBID. Some settings are pushed down
to a new parm file. There is also a small amount of setting
shuffling between parm files.

Also updated the python script to conform to bash variables more
and provide better error messages. This includes using mail.py to
email when it fails to produce an increment for warm start to warn
that the forecast may be degraded.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266