Convert atmos_prep to a single MPMD job
At the suggestion of NCO, the atmos_prep metatask has been converted
to a single MPMD job since recenter relies on all prep jobs to be
run. This change makes it easier for NCO to monitor and control job
flow.

To execute this change, most of the existing atmos_prep script was
pushed down to a new script at the ush level, which the shell of
the former script is filled with the creation of the MPMD command
file.

So that the correct resources are requested (currently 1 node/member),
a new function calculaing the needed resources is added to XML_For_
Tasks. Also brought most of that script up to the PEP8 standard, if
only because my python linter flags all the lines with errors and it
can be quite distracting.

Refs: #63941, #76266