Clean j-job and error messages for forecast
Remakes the j-job to NCO standards. Places where jid ($$) were
used are replaced with $LSB_JOBID. Much of the code from the
former j-job is pushed down to a new script-level script. That
script calls the script-level global workflow script within
global workflow.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266