Clean j-job and error messages for post
Remakes the j-job to NCO standards. Places where jid ($$) were used are
replaced with $LSB_JOBID. Much of the code from the former j-job is
pushed down to a new script-level script. The previous script-level
script is pushed down to ush-level.

Formerly, there were large blocks of nearly identical code for analysis
and forecast hours that if control and output files already existed.
Replaced these with a function check_existing_post to eliminate need for
redundant code.

The ush/gefs_restart_cleanup.sh script and the section that called it
when $RERUN = YES were broken. The script is deleted; in its place,
while we are checking control files anyway in check_existing_post, the
control files are deleted for $RERUN = YES.

Additionally, a lot of output and error checking is standardized,
including the PS4 and error messages.

Refs: #63941, #76266